Reduce DOM queries and layout thrashing by caching board size
In my current quest to reduce unnecessary style recalculations and
layouting, I saw that the sidebar resizer forces style recalculations by
querying the board height at least whenever the current encounter turn
changes and in many other cases. Fortunately, caching the board height
is trivial and save with help of a resize observer.
This in my case reduces the ui freeze on combat turn progression and
when adding effects to actors by 20ms.